+++ src/marsyas/PeakLabeler.cpp
@@ -64,7 +64,7 @@ PeakLabeler::myProcess(realvec& in, realvec& out)
 	
 	const realvec& peakLabels = ctrl_peakLabels_->to<mrs_realvec>(); //reading unlocked and linked control!!! --> should use MarControlAccessor [TODO][!]
 
-	if(peakLabels.getSize() != outPeakView.getTotalNumPeaks())
+	if(peakLabels.getSize() != (size_t)outPeakView.getTotalNumPeaks())
 	{
 		MRSERR("PeakLabeler::myProcess - peakLabels control and input peaks number mismatch! Labeling not performed!");
 	}
